Fetch developer dir dist info from intermediate table when date range is not selected
This is a temp solution.
The developer dir dist in result table contains time series info, so sum
of these numbers will contain duplicated developers.

For now, if date range is not set, fetch data from the intermediate
table(without time series column), so the number will (roughly) match with
timezone dist. While the date range is not applied to timezone dist